refactor: Move paper queries in own file
