+++ libs/cozy/paperCipher.ts
@@ -1,6 +1,5 @@
 /* eslint-disable no-console */
 // Cozy customization
-import { Q } from "cozy-client";
 import CozyClient from "cozy-client/types/CozyClient";
 
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
@@ -12,96 +11,7 @@ import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 
 import { convertNoteToCipherResponse, isNote, fetchNoteIllustrationUrl } from "./note.helper";
 import { convertPaperToCipherResponse } from "./paper.helper";
-
-const fetchPapers = async (client: CozyClient) => {
-  const filesQueryByLabels = buildFilesQueryWithQualificationLabel();
-
-  const data = await client.queryAll(filesQueryByLabels.definition(), filesQueryByLabels.options);
-
-  const hydratedData = client.hydrateDocuments("io.cozy.files", data);
-
-  return hydratedData;
-};
-
-const fetchPaper = async (client: CozyClient, _id: string) => {
-  const fileQueryWithContact = buildFileQueryWithContacts(_id);
-
-  const { data } = await client.query(
-    fileQueryWithContact.definition(),
-    fileQueryWithContact.options
-  );
-
-  return data[0];
-};
-
-export const buildFilesQueryWithQualificationLabel = () => {
-  const select = [
-    "name",
-    "mime",
-    "referenced_by",
-    "metadata.country",
-    "metadata.datetime",
-    "metadata.expirationDate",
-    "metadata.noticePeriod",
-    "metadata.qualification.label",
-    "metadata.referencedDate",
-    "metadata.number",
-    "metadata.contractType",
-    "metadata.refTaxIncome",
-    "metadata.title",
-    "metadata.AObtentionDate",
-    "metadata.BObtentionDate",
-    "metadata.CObtentionDate",
-    "metadata.DObtentionDate",
-    "metadata.page",
-    "metadata.version",
-    "cozyMetadata.createdByApp",
-    "cozyMetadata.sourceAccountIdentifier",
-    "cozyMetadata.favorite",
-    "created_at",
-    "dir_id",
-    "updated_at",
-    "type",
-    "trashed",
-  ];
-
-  return {
-    definition: () =>
-      Q("io.cozy.files")
-        .where({
-          type: "file",
-          trashed: false,
-        })
-        .partialIndex({
-          "metadata.qualification.label": {
-            $exists: true,
-          },
-          "cozyMetadata.createdByApp": { $exists: true },
-        })
-        .select(select)
-        .limitBy(1000)
-        .include(["contacts"])
-        .indexFields(["type", "trashed"]),
-    options: {
-      as: `io.cozy.files/metadata_qualification_label`,
-    },
-  };
-};
-
-export const buildFileQueryWithContacts = (_id: string) => {
-  return {
-    definition: () =>
-      Q("io.cozy.files")
-        .where({
-          _id: _id,
-        })
-        .limitBy(1)
-        .include(["contacts"]),
-    options: {
-      as: `io.cozy.files/byIdWithContacts`,
-    },
-  };
-};
+import { fetchPapers, fetchPaper } from "./queries";
 
 const convertPapersAsCiphers = async (
   cipherService: any,
